Find current waypoint
Route now has a static method for finding the "current"
waypoint ("pending" waypoint with minumum date).

My first instinct was to have this as an instance method on the Route
class, but this doesn't work because TypeScript doesn't seem to cast
JSON objects to a class deeply and provide methods. There's probably a
way to get it to do this, and if there's a best practice I can revisit
this. But for now, using a static method because at least it 1) allows
typechecking and 2) is functional, at least.